fix: Op::Call/Apply must use u16 arg count, not u8
The bytecode decoder truncated the 2-byte Call arg count to u8 via
as-u8, silently wrapping values >= 256 to 0. This caused
dired-mode-map defvar-keymap (which has ~308 args to define-keymap)
to decode as Call(0) instead of Call(308).

With Call(0), the VM took the last pushed constant
(dired-enable-click-to-select-mode) as the function to call, instead
of correctly popping 308 args and calling define-keymap below them.

GNU Emacs Bcall uses a 2-byte encoding (opcode 39 = Bcall + FETCH2)
supporting up to 65535 args. Now Op::Call and Op::Apply use u16.

Fixes: void-function dired-enable-click-to-select-mode when loading
dired.elc, which cascaded to break magit and all dired-dependent
packages.

+++ neovm-core/src/emacs_core/bytecode/compiler.rs
@@ -281,7 +281,7 @@ impl Compiler {
             for arg in tail {
                 self.compile_expr(func, arg, true);
             }
-            self.emit_tracked(func, Op::Call(tail.len() as u8));
+            self.emit_tracked(func, Op::Call(tail.len() as u16));
 
             if !for_value {
                 self.emit_tracked(func, Op::Pop);
@@ -297,7 +297,7 @@ impl Compiler {
                     for arg in tail {
                         self.compile_expr(func, arg, true);
                     }
-                    self.emit_tracked(func, Op::Call(tail.len() as u8));
+                    self.emit_tracked(func, Op::Call(tail.len() as u16));
                     if !for_value {
                         self.emit_tracked(func, Op::Pop);
                     }
@@ -311,7 +311,7 @@ impl Compiler {
         for arg in tail {
             self.compile_expr(func, arg, true);
         }
-        self.emit_tracked(func, Op::Call(tail.len() as u8));
+        self.emit_tracked(func, Op::Call(tail.len() as u16));
         if !for_value {
             self.emit_tracked(func, Op::Pop);
         }
@@ -418,7 +418,7 @@ impl Compiler {
                     for arg in &tail[1..] {
                         self.compile_expr(func, arg, true);
                     }
-                    self.emit_tracked(func, Op::Call(tail.len().saturating_sub(1) as u8));
+                    self.emit_tracked(func, Op::Call(tail.len().saturating_sub(1) as u16));
                     if !for_value {
                         self.emit_tracked(func, Op::Pop);
                     }
